refactor: HidePRsByTitleMenu と actions column helper を専用モジュールに分離
pr-block.tsx は PR block の視覚化が主題だが、HidePRsByTitleMenu と
PRHideByTitleFilterContext は 5+ ファイルから参照される汎用部品になっていた。
また 4 つの +columns.tsx で「isAdmin なら actions 列を push」するブロックが
重複していた。

- hide-prs-by-title-menu.tsx を新設し、menu / context / hidePrActionsColumn
  helper を集約
- throughput × 3 と feedback-columns で helper を spread 適用
- pr-drill-down-sheet.tsx の className={reviewTime ? '' : 'ml-auto'} を
  親 flex container で ml-auto クラスタにまとめて解消
