+++ app/routes/$orgSlug/analysis/reviews/+components/pr-drill-down-sheet.tsx
@@ -10,7 +10,7 @@ import {
   SheetTitle,
 } from '~/app/components/ui/sheet'
 import { parseRiskAreas } from '~/app/libs/pr-classify'
-import { HidePRsByTitleMenu } from '~/app/routes/$orgSlug/+components/pr-block'
+import { HidePRsByTitleMenu } from '~/app/routes/$orgSlug/+components/hide-prs-by-title-menu'
 
 export interface DrillDownPR {
   number: number
@@ -85,19 +85,14 @@ export function PRDrillDownSheet({
                     </Avatar>
                     {pr.authorDisplayName ?? pr.author}
                   </span>
-                  {pr.reviewTime != null && pr.reviewTime > 0 && (
-                    <span className="text-muted-foreground ml-auto text-xs">
-                      {formatHours(pr.reviewTime * 24)}
-                    </span>
-                  )}
-                  <HidePRsByTitleMenu
-                    title={pr.title}
-                    className={
-                      pr.reviewTime != null && pr.reviewTime > 0
-                        ? ''
-                        : 'ml-auto'
-                    }
-                  />
+                  <div className="ml-auto flex items-center gap-2">
+                    {pr.reviewTime != null && pr.reviewTime > 0 && (
+                      <span className="text-muted-foreground text-xs">
+                        {formatHours(pr.reviewTime * 24)}
+                      </span>
+                    )}
+                    <HidePRsByTitleMenu title={pr.title} />
+                  </div>
                 </div>
                 <a
                   href={pr.url}
