+++ app/routes/$orgSlug/throughput/deployed/+columns.tsx
@@ -5,7 +5,7 @@ import { Badge, HStack } from '~/app/components/ui'
 import { Avatar, AvatarFallback, AvatarImage } from '~/app/components/ui/avatar'
 import dayjs from '~/app/libs/dayjs'
 import { complexitySortingFn } from '~/app/libs/pr-classify'
-import { HidePRsByTitleMenu } from '~/app/routes/$orgSlug/+components/pr-block'
+import { hidePrActionsColumn } from '~/app/routes/$orgSlug/+components/hide-prs-by-title-menu'
 import type { PullRequest } from './index'
 
 export function createColumns(
@@ -13,7 +13,7 @@ export function createColumns(
   orgSlug: string,
   isAdmin: boolean,
 ): ColumnDef<PullRequest>[] {
-  const columns: ColumnDef<PullRequest>[] = [
+  return [
     {
       accessorKey: 'author',
       header: ({ column }) => (
@@ -173,20 +173,6 @@ export function createColumns(
       ),
       enableHiding: false,
     },
+    ...hidePrActionsColumn<PullRequest>(isAdmin, (r) => r.title),
   ]
-
-  if (isAdmin) {
-    columns.push({
-      id: 'actions',
-      header: '',
-      cell: ({ row }) =>
-        row.original.title ? (
-          <HidePRsByTitleMenu title={row.original.title} />
-        ) : null,
-      enableHiding: false,
-      enableSorting: false,
-    })
-  }
-
-  return columns
 }
